pyproject.toml
[project]
name = "anton"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pydantic-settings>=2.8.1",
    "python-docx>=1.1.2",
    "rich>=14.0.0",
    "typer>=0.15.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ipdb>=0.13.13",
    "ipython>=9.1.0",
    "isort>=6.0.1",
]
